requirements.txt
PyQt6>=6.5.0
PyQt6-Qt6>=6.5.0
PyQt6-sip>=13.5.0
matplotlib>=3.7.0
tensorflow>=2.12.0
keras>=2.12.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
h5py>=3.8.0
tqdm>=4.65.0
scapy>=2.5.0
dpkt>=1.9.8
pyshark>=0.5.3
geoip2>=4.7.0
python-whois>=0.8.0
ipaddress>=1.0.23
python-registry>=1.3.1
olefile>=0.46
rarfile>=4.1
py7zr>=0.20.8
psutil>=5.9.0
Pillow>=9.5.0
opencv-python>=4.8.0
pycryptodome>=3.19.0
cryptography>=41.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
python-pptx>=0.6.21
PyPDF2>=3.0.0
pdfplumber>=0.10.0
mutagen>=1.46.0
hachoir>=3.2.0
pypiwin32>=223; sys_platform == 'win32'
pywin32>=306; sys_platform == 'win32'
Jinja2>=3.1.0
reportlab>=4.0.0
WeasyPrint>=60.0
markdown>=3.4.0
PyYAML>=6.0
loguru>=0.7.0
colorama>=0.4.6
python-dateutil>=2.8.0
pytz>=2023.3
click>=8.1.0
sqlalchemy>=2.0.0
pyinstaller>=5.13.0